+++ actix-http/src/header/shared/quality.rs
@@ -38,8 +38,11 @@ impl Quality {
     /// The maximum quality value, equivalent to `q=1.0`.
     pub const MAX: Quality = Quality(MAX_QUALITY_INT);
 
-    /// The minimum quality value, equivalent to `q=0.0`.
-    pub const MIN: Quality = Quality(0);
+    /// The minimum, non-zero quality value, equivalent to `q=0.001`.
+    pub const MIN: Quality = Quality(1);
+
+    /// The zero quality value, equivalent to `q=0.0`.
+    pub const ZERO: Quality = Quality(0);
 
     /// Converts a float in the range 0.0–1.0 to a `Quality`.
     ///
@@ -51,7 +54,7 @@ impl Quality {
         // Check that `value` is within range should be done before calling this method.
         // Just in case, this debug_assert should catch if we were forgetful.
         debug_assert!(
-            (0.0f32..=1.0f32).contains(&value),
+            (0.0..=MAX_QUALITY_FLOAT).contains(&value),
             "q value must be between 0.0 and 1.0"
         );
 
@@ -185,6 +188,10 @@ mod tests {
 
     #[test]
     fn display_output() {
+        assert_eq!(Quality::ZERO.to_string(), "0");
+        assert_eq!(Quality::MIN.to_string(), "0.001");
+        assert_eq!(Quality::MAX.to_string(), "1");
+
         assert_eq!(q(0.0).to_string(), "0");
         assert_eq!(q(1.0).to_string(), "1");
         assert_eq!(q(0.001).to_string(), "0.001");